backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
pydantic==2.8.0
google-generativeai
pymupdf==1.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
jinja2==3.1.4
weasyprint==62.3
sendgrid==6.11.0
python-multipart==0.0.9
httpx==0.27.0
langgraph==0.0.30
langchain-google-genai
reportlab==4.1.0
